backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
llama-index = "^0.10.15"
llama-index-core = "^0.10.15"
llama-index-agent-openai = "0.1.5"
werkzeug = "^3.0.3"
python-multipart = "^0.0.9"
nest-asyncio = "^1.6.0"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"